requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-groq==0.2.0
langchain-huggingface==0.1.0
chromadb==0.5.3
fastapi==0.112.0
uvicorn==0.30.0
pydantic==2.8.2
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
sentence-transformers==3.0.1
huggingface-hub==0.24.0